Fixed the bug using a promise.
This implementation creates an array of promises by mapping each city to a call to getData, and then waits for all the promises to resolve using Promise.all. Once all the promises are resolved, the resulting data array is returned.

This code waits for the data to be fetched before calling setDefaultCities with the resulting array. This should cause the component to re-render with the updated state.